requirements.txt
Django==4.0.4; python_version >= '3.8'
Django==3.2.18; python_version < '3.8'
django-environ==0.10.0
psycopg2-binary==2.9.6
google-cloud-secret-manager==2.16.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
firebase-admin==6.1.0
keras
tensorflow